requirements.txt
Flask==2.0.1
Werkzeug==2.0.3
SQLAlchemy==1.4.23
Flask-SQLAlchemy==2.5.1
python-dotenv==0.19.0
PyMySQL==1.0.2 